Move BrentFocusOptimizer back ingo OughtaFocus
This is necessary for the build to work, for now.
